Removed cached list check
If a tournament list updates after it has been first submitted (see
3900) then the current code will not update it on the client's local
storage.  This means that some number of lists will probably be updated
before they age out of the cache, but never display correctly if they
were originally submitted without lists in them.

This is a stop-gap to make sure that all client-side cached tournament
information will be ignored in favor of gh-pages versions, which are
updated every four hours (for new lists) and every day (for old,
possibly recently-updated info).

A better option would be to figure out a way to save a "dirty" flag on
the Linux server for any tournament lists that have been updated
recently so that XWS-Bench can delete the cached version from the client
and replace it with the updated version.  But since the performance hit
for fully refreshing the client-side listings every time is minimal,
while the development outlay for implementing a full cache invalidation
scheme over two servers and the client is nontrivial, I am not going to
put any further effort into it.